refactor: decompose chromium-debug-use.mjs into modules to reduce complexity
